Bug fix: Cheating would be detected in gone loud and mod is stealth only
Whenever the player drops the last bag in the stack, the system would
detect cheating and would make carry disappear. This would only
happen if the player had the mod configured to be used only during
stealth, and carrying a bag before the alarm went off.

After the alarm went off, and cheating had been detected once (hence,
all bags had been dropped), picking up a bag and dropping it would also
be detected as cheating. However, the bag would not disappear.

Now, when the player is dropping a bag, master drop_carry is called
before BLT_CarryStacker:RemoveCarry. This makes the mod's state be
updated after the master call, so when the anticheat system requests
to verify_carry, the mod would still be enabled.