Changed the way the mod's state is handled
Before, the mod only had two states: Enabled and Disabled. This is, the
state was just a boolean value indicating whether the mod was enabled
or not. However, there is an edge case: what if the mod is to
transition into the Disabled state but the player is still carrying
more than one bag?

Before, this would be handled by still enabling the
mod if the player was carrying bags (and the mod is configured to be
used on stealth only and its loud). However, this means we have to
add into PlayerManager:can_carry new logic to differentiate between the
actual Enabled state and the transition from Enabled to Disabled.

The way it is currently implemented, we have added a new intermediate
state: Being_disabled. Now, e just need to check the mod's state,
instead of adding this new logic into PlayerManager:can_carry.
Moreover, with this change in the way state is handled, a minor bug has
been fixed:
The player used to be shown shown the help message to carry a new bag
after the alarm went off and already carrying a bag (having the mod's
stealth only option selected).

Moreover, this change in the way state is handled will allow
introducing a new option to the game: A general enable/disable option.
This is, while in-game, the player would be able to toggle whether the
mod is being used or not. As of now, the player has to unload the mod
from BLT mods and restart the game. So not only it cannot be
enabled/disabled manually while in a heist, it also implies the player
has to restart the game. This new option will soon be introduced! :)

Currently, the mod's state is calculated on-the-fly in
BLT_CarryStacker:GetModState. However, this might change it in the
future so that the state is updated based in a callback system. This
is, for example, when the alarm is raised, the state will be changed to
Being_disabled (if stealth only). That way, the state will not have to
be calculated every time it is needed.