Bug fix: it mod was to be disabled but #stack > 0 then cheats detected
Before, the stack was only checked to be empty when changing to the
DISABLED state in the stealth only option. Now, in general, the mod
cannot transition into the DISABLED state if #stack > 0.

This fixes the bug in which the player was online, a having the offline
only toggle set to false. Then, they pick a bag and set the offline
only toggle to true. Afterwards, when trying to throw the bag, the game
would detect cheating and the bag would disappear (Same as in 81ce4611506b9c741ebab976f97380768292300c).