Added chat message to inform the mod is being disabled
While in the BEING_DISABLED state (for example if the mod is
stealth only, the player is carrying bags and the alarm went off), the
player is still able to use the mod (but not pick more bags). It will
not be until the player drops the bags that the mod will be disabled
completely. This situation is now informed to the player by a chat
message only visible to them.